Move AssetParams to its own class.
